requirements.txt
gdown==4.5.1
pytorch_lightning==1.6.5
omegaconf==2.2.2
einops==0.4.1
torchmetrics==0.9.2
wandb==0.12.21
lightning-bolts==0.5.0
albumentations==1.2.1
timm==0.4.12
ttach==0.0.3
segmentation_models_pytorch==0.2.1